rename rtu->modbusrtu packages
