+++ modbusrtu/state.go
@@ -1,4 +1,4 @@
-package rtu
+package modbusrtu
 
 import (
 	"fmt"
